requirements.txt
resend
python-dotenv
requests